translator: refactor storage-format implementation into own translator
Refactoring the existing translator to split the implementation between
a base translator (`ConfluenceBaseTranslator`) and a new storage
format-specific translator (`ConfluenceStorageFormatTranslator`). This
is to help prepare for prospect changes where a new translator aims to
support the Atlassian Document format.
